tools/ppcdis/requirements.txt
colorama
capstone
pyelftools
pylibyaml
PyYAML
